Web/Controllers/HomeController.cs
Modified to provide the list of a company's orders to the view directly as opposed to an ajax call from the view itself.

Web/Infrastructure/Database.cs

Modifed to better handle management of the active sql connection. Additionally modified ExecuteReader and ExecuteNonQuery to open and dispose of both the SqlConnection and the SqlCommand provided.

Web/Infrastructure/OrderService.cs

Updated the GetOrdersForCompany function to account for the modified DataBase.cs file.

Web/Views/Home/Index.cshtml

Modified the view to bind to the Order model. The view can render the orders server side and send the full page to the client.